ROFO-60 refactor: 검색 기록 기능들을 search 패키지로 이동
